+++ LinearStackCompressor.h
@@ -197,7 +197,7 @@ protected:
 
     virtual void writeHeader();
     virtual std::string headerDescription();
-    virtual inline unsigned long idCode () {
+    virtual inline uint32_t idCode () {
         return 0xa3d2d45d; //CRC32 hash of "LinearStackCompressor" from fileformat.info
     }
 private:
