+++ NameValueMetaData.h
@@ -18,13 +18,13 @@
 
 class NameValueMetaData : public ImageMetaData{
 public:
-    static const unsigned long IdCode = 0xc15ac674; //CRC32 hash of "NameValueMetaData" from http://www.fileformat.info/tool/hash.htm?text=NameValueMetaData
+    static const uint32_t IdCode = 0xc15ac674; //CRC32 hash of "NameValueMetaData" from http://www.fileformat.info/tool/hash.htm?text=NameValueMetaData
     NameValueMetaData();
    virtual ~NameValueMetaData();
     virtual void toDisk (std::ofstream &os) const;
     virtual std::string saveDescription ()const;
     virtual int sizeOnDisk ()const;
-    virtual unsigned long idCode()const {
+    virtual uint32_t idCode()const {
         return IdCode;
     }
     virtual std::map<std::string, double> getFieldNamesAndValues(void)const;
